% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aiptw.R
\name{aiptw}
\alias{aiptw}
\title{Augmented Inverse Probability Treatment Weighted (A-IPTW) Estimator}
\usage{
aiptw(abar, a, g_a, Q_aw, y)
}
\arguments{
\item{abar}{Treatment of interest}

\item{a}{Indicator of treatment assignment}

\item{g_a}{Estimate of g(A|W)}

\item{Q_aw}{Estimate of Q(Y|A,W)}

\item{y}{Indicator of outcome observed}
}
\value{
\code{aiptw} returns the A-IPTW estimate.
}
\description{
\code{aiptw} Returns the estimate of E Y_a using the efficient estimator developed by Robins et al.
}
\details{
Requires that both g_a and Q_aw be fit manually outside of the function.
}

