% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateData}
\alias{generateData}
\title{Generating data}
\usage{
generateData(n, Q_W, g_A, Q_Y, abar = NULL)
}
\arguments{
\item{n}{Number of observations to generate.}

\item{Q_W}{Function for generating W.}

\item{g_A}{Function for generating A.}

\item{Q_Y}{Function for generating Y.}

\item{abar}{Whether to generate a post-intervention distribution (and if so, the specified one), or to just generate the observed distribution.}
}
\value{
\code{generateData} returns a \code{data.frame} object, which contains realized observations from the user specified distribution.
}
\description{
\code{generateData} Generates data based on a user specified distribution.
}
\details{
Details
}

