% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iptw.R
\name{iptw}
\alias{iptw}
\title{Inverse Probability Treatment Weighted (IPTW) Estimator}
\usage{
iptw(abar, a, g_a, y)
}
\arguments{
\item{abar}{Treatment of interest}

\item{a}{Indicator of treatment assignment}

\item{g_a}{Estimate of g(A|W)}

\item{y}{Indicator of outcome observed}
}
\value{
\code{iptw} returns the IPTW estimate.
}
\description{
\code{iptw} Returns the estimate of E Y_a using the estimator popularized by Robins et al.
}
\details{
Requires that g_a be fit manually outside of the function.
}

