% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{P_a}
\alias{P_a}
\title{Simulation post-intervention data.}
\description{
\code{P_a} contains the generated true post-intervention distribution based on 1 million observations using the true distribution.
}
\author{
Linh Tran \email{tranlm@berkeey.edu}
}
\references{
\url{data_blah.com}
}
\keyword{data}

