% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Simulation resulting estimates.}
\description{
\code{results} contains the simulation resulting estimates from each estimator considered.
}
\author{
Linh Tran \email{tranlm@berkeey.edu}
}
\references{
\url{data_blah.com}
}
\keyword{data}

